mdm模型增删查改api优化

+++ fisk-mdm-model/fisk-mdm-model-web/src/main/java/com/fisk/mdm/service/impl/ModelServiceImpl.java
@@ -1,5 +1,7 @@
 package com.fisk.mdm.service.impl;
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.fisk.common.core.response.ResultEntity;
@@ -14,7 +16,10 @@ import com.fisk.mdm.dto.model.ModelQueryDTO;
 import com.fisk.mdm.vo.model.ModelVO;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service()
 public class ModelServiceImpl extends ServiceImpl<ModelMapper, ModelPO> implements IModelService {
@@ -40,6 +45,14 @@ public class ModelServiceImpl extends ServiceImpl<ModelMapper, ModelPO> implemen
      */
     @Override
     public ResultEnum addData(ModelDTO model) {
+        //根据name查询，若已存在则不添加
+        QueryWrapper<ModelPO> wrapper = new QueryWrapper<>();
+        wrapper.eq("name",model.name);
+        ModelPO modelPO = baseMapper.selectOne(wrapper);
+        if(modelPO != null){
+            return ResultEnum.DATA_EXISTS;
+        }
+
         boolean isInsert = baseMapper.insert(ModelMap.INSTANCES.dtoToPo(model)) > 0;
         if (!isInsert) {
             return ResultEnum.SAVE_DATA_ERROR;
@@ -54,7 +67,17 @@ public class ModelServiceImpl extends ServiceImpl<ModelMapper, ModelPO> implemen
      */
     @Override
     public ResultEnum editData(ModelDTO model) {
-        boolean isUpData = baseMapper.updateById(ModelMap.INSTANCES.dtoToPo(model)) > 0;
+        ModelPO modelPO = baseMapper.selectById(model.id);
+
+        //判断数据是否存在
+        if (modelPO==null){
+            return ResultEnum.DATA_NOTEXISTS;
+        }
+
+        //把DTO转化到查询出来的PO上
+        modelPO = ModelMap.INSTANCES.dtoToPo(model);
+
+        boolean isUpData = baseMapper.updateById(modelPO) > 0;
         if (!isUpData) {
             return ResultEnum.UPDATE_DATA_ERROR;
         }
@@ -68,6 +91,11 @@ public class ModelServiceImpl extends ServiceImpl<ModelMapper, ModelPO> implemen
      */
     @Override
     public ResultEnum deleteDataById(Integer id) {
+        //判断数据是否存在
+        if (baseMapper.selectById(id)==null){
+            return ResultEnum.DATA_NOTEXISTS;
+        }
+
         boolean isDelete = baseMapper.deleteById(id) > 0;
         if (!isDelete) {
             return ResultEnum.DATA_NOTEXISTS;
@@ -82,6 +110,9 @@ public class ModelServiceImpl extends ServiceImpl<ModelMapper, ModelPO> implemen
      */
     @Override
     public Page<ModelVO> getAll(ModelQueryDTO query) {
-        return baseMapper.getAll(query.page, query);
+
+        Page<ModelVO> all = baseMapper.getAll(query.page, query);
+
+        return all;
     }
 }
